The new TTableDescription object is a TObject based table description that contains all information about a table. It is attached to each tree using the tree's UserInfo. To read this information from a root file the new libroottable.so needs to be loaded. However, the tree itself can also be used without this library.
The disadvantage is that ROOT will spit out an error message for each tree
if libroottable.so is not loaded. The advantage is that this opens the 
potential to read data from root files using tableio and makes the use of 
the tabletranscriber possible.

WARNING: This is highly experimental because I have no experience with some
of the stuff I am using here. Lots of testing needed before a release.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/rootwriter/trunk@68014 16731396-06f5-0310-8873-f7f720988828